apps/python-server/requirements.txt
faster-whisper
resemblyzer
webrtcvad
numpy
scipy
scikit-learn
pycountry
librosa
fastapi[standard]
dotenv
requests
uvicorn
python-multipart
starlette
sqlalchemy
ffmpeg-python